chore: 🔧 update package dependencies and configurations
This commit updates the `pnpm` version in `package.json` to `10.15.0` and reflects the changes in `pnpm-lock.yaml`. It also modifies the dependency name for `@gemini-proxy/vercel` to `@lehuygiang28/gemini-proxy-vercel` in multiple files. Additionally, it adds `tsup` as a dependency with version `8.5.0` in relevant packages and enhances TypeScript configurations by enabling `downlevelIteration` and specifying `types` for better type support. These changes aim to improve project consistency and compatibility.